:recycle: move image_app.models mixins in a dedicated module
